requirements.txt
torch
torchvision
scikit-learn
numpy
scipy
matplotlib
jupyter
tensorboardX
tqdm
pyyaml
pandas
lightning
click
termcolor
